refactor: unify DocumentProcessor
